script/requirements.txt
-i https://pypi.org/simple
certifi==2023.7.22 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
clang==16.0.1.1
contourpy==1.1.0 ; python_version >= '3.8'
cryptography==41.0.3
cycler==0.11.0 ; python_version >= '3.6'
deprecated==1.2.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
et-xmlfile==1.1.0 ; python_version >= '3.6'
exceptiongroup==1.1.3 ; python_version < '3.11'
feather-format==0.4.1
fonttools==4.42.1 ; python_version >= '3.8'
ftfy==6.1.1
gitdb==4.0.10 ; python_version >= '3.7'
gitpython==3.1.32
idna==3.4 ; python_version >= '3.5'
importlib-resources==6.0.1 ; python_version < '3.10'
iniconfig==2.0.0 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
kiwisolver==1.4.4 ; python_version >= '3.7'
libclang==16.0.6
lizard==1.17.10
lxml==4.9.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==2.1.3 ; python_version >= '3.7'
matplotlib==3.7.2
networkx==3.1
numpy==1.25.2 ; python_version >= '3.9'
openpyxl==3.1.2
packaging==23.1 ; python_version >= '3.7'
pandas==2.0.3
pillow==10.0.0 ; python_version >= '3.8'
pluggy==1.2.0 ; python_version >= '3.7'
psycopg2-binary==2.9.7
pyarrow==12.0.1 ; python_version >= '3.7'
pycparser==2.21
pydriller==2.5.1
pygithub==1.59.1
pyjwt[crypto]==2.8.0 ; python_version >= '3.7'
pynacl==1.5.0 ; python_version >= '3.6'
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
pytest==7.4.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
pyyaml==6.0.1
requests==2.31.0 ; python_version >= '3.7'
sarif-tools==1.0.0
scipy==1.11.2
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.0 ; python_version >= '3.6'
tabulate==0.9.0
tomli==2.0.1 ; python_version < '3.11'
types-pytz==2023.3.0.1
tzdata==2023.3 ; python_version >= '2'
urllib3==2.0.4
wcwidth==0.2.6
whatthepatch==1.0.5
wrapt==1.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
zipp==3.16.2 ; python_version < '3.10'
